pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "anima-nasdetr"
version = "0.1.0"
description = "Paper-faithful NAS-DETR scaffold for sonar object detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch>=2.2",
  "numpy>=1.26",
  "pillow>=10.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/anima_nasdetr"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
